Refactoring.

+++ modules/db/src/main/scala/net/pointsgame/db/repositories/RepositoryBase.scala
@@ -13,16 +13,16 @@ abstract class RepositoryBase[T <: Entity, B <: BaseTable[T]](db: Database) exte
   override def all: Task[Seq[T]] = db.run {
     query.result
   }.asScalaz
-  override def exists(id: Int): Task[Boolean] = db.run {
+  override def exists(id: Long): Task[Boolean] = db.run {
     query.filter(_.id === id).exists.result
   }.asScalaz
-  override def getById(id: Int): Task[Option[T]] = db.run {
+  override def getById(id: Long): Task[Option[T]] = db.run {
     query.filter(_.id === id).take(1).result.map(_.headOption)
   }.asScalaz
-  override def deleteById(id: Int): Task[Boolean] = db.run {
+  override def deleteById(id: Long): Task[Boolean] = db.run {
     query.filter(_.id === id).take(1).delete.map(_ > 0)
   }.asScalaz
-  override def insert(entity: T): Task[Int] = db.run {
+  override def insert(entity: T): Task[Long] = db.run {
     assert(entity.isNew)
     query.returning(query.map(_.id)) += entity
   }.asScalaz
