+++ modules/db/src/main/scala/net/pointsgame/db/schema/BaseTable.scala
@@ -3,5 +3,5 @@ package net.pointsgame.db.schema
 import slick.driver.SQLiteDriver.api._
 
 abstract class BaseTable[T](tag: Tag, name: String) extends Table[T](tag, name) {
-  def id = column[Int]("Id", O.PrimaryKey, O.AutoInc)
+  def id = column[Long]("Id", O.PrimaryKey, O.AutoInc)
 }
