+++ modules/domain/src/main/scala/net/pointsgame/domain/model/Entity.scala
@@ -1,7 +1,7 @@
 package net.pointsgame.domain.model
 
 trait Entity {
-  def id: Option[Int]
+  def id: Option[Long]
   def isNew: Boolean =
     id.isEmpty
 }
