+++ modules/domain/src/main/scala/net/pointsgame/domain/model/Room.scala
@@ -1,5 +1,5 @@
 package net.pointsgame.domain.model
 
-case class Room(id: Option[Int], name: String) extends Entity {
+case class Room(id: Option[Long], name: String) extends Entity {
   assert(name.nonEmpty)
 }
