+++ modules/domain/src/main/scala/net/pointsgame/domain/model/RoomMessage.scala
@@ -2,6 +2,6 @@ package net.pointsgame.domain.model
 
 import com.github.nscala_time.time.Imports._
 
-case class RoomMessage(id: Option[Int], body: String, roomId: Int, senderId: Int, sendingDate: DateTime) extends Entity {
+case class RoomMessage(id: Option[Long], body: String, roomId: Long, senderId: Long, sendingDate: DateTime) extends Entity {
   assert(body.nonEmpty)
 }
