+++ modules/domain/src/main/scala/net/pointsgame/domain/services/TokenService.scala
@@ -2,7 +2,6 @@ package net.pointsgame.domain.services
 
 import com.github.nscala_time.time.Imports._
 import scalaz._
-import Scalaz._
 import scalaz.concurrent.Task
 import net.pointsgame.domain.model.Token
 import net.pointsgame.domain.repositories.TokenRepository
@@ -10,7 +9,7 @@ import net.pointsgame.domain.helpers.Tokenizer
 import net.pointsgame.domain.{ DomainException, Constants }
 
 final case class TokenService(tokenRepository: TokenRepository) {
-  def create(userId: Int): Task[Token] = {
+  def create(userId: Long): Task[Token] = {
     val now = DateTime.now()
     val token = Token(None, userId, Tokenizer.generate(Constants.tokenLength), now, now, false)
     tokenRepository.insert(token).map(id => token.copy(id = Some(id)))
